% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_CI_par.R
\name{coef_CI_par}
\alias{coef_CI_par}
\title{Confidence Intervals and Estimates of Each Regression Coefficient - Parallelized}
\usage{
coef_CI_par(lrbs, alpha = 0.05)
}
\arguments{
\item{lrbs}{A linear_reg_bs or linear_reg_bs_par object containing BLB regression
coefficient estimates.}

\item{alpha}{The significance level. Default value is 0.05.}
}
\value{
The overall confidence interval for each regression coefficient, along with
its overall estimate.
}
\description{
This function takes in a list of linear regression coefficient estimates generated
by a Bag of Little Bootstraps procedure. Then, empirical confidence intervals and
point estimates of each coefficient are determined for each subsample. Afterwards,
the endpoints of all confidence intervals are averaged to form overall confidence
intervals, and point estimates are averaged to form overall estimates. It should be
noted that the confidence intervals are not multiple confidence intervals. For
Bonferroni-corrected confidence intervals, divide the desired value of alpha by the
number of regression coefficients. The difference between this function and coef_CI
is that this function uses parallel processing through furrr's future_map function.
}
