% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI_par.R
\name{PI_par}
\alias{PI_par}
\title{Prediction Intervals and Estimates for New Data - Parallelized}
\usage{
PI_par(lrbs, x, alpha = 0.05)
}
\arguments{
\item{lrbs}{A linear_reg_bs or linear_reg_bs_par object containing BLB regression
coefficient estimates.}

\item{x}{A dataframe of the explanatory variables of unseen observations.}

\item{alpha}{The significance level. Default value is 0.05.}
}
\value{
The prediction intervals and estimates for the response variable of each
unseen observation.
}
\description{
This function takes in a list of linear regression coefficient estimates generated
by a Bag of Little Bootstraps procedure, and a dataframe of observations without the
response variable. The response variable for each observation is predicted using
each vector of coefficient estimates for each sample. Then, empirical prediction
intervals and point estimates for the response variable of each observation are
determined for each sample. Afterwards, the endpoints of all intervals are averaged
to form overall prediction intervals, and point estimates are averaged to form
overall predictions. It should be noted that the prediction intervals are not
multiple prediction intervals. For Bonferroni-corrected prediction intervals, divide
the desired value of alpha by the number of observations. The difference between
this function and PI is that this function uses parallel processing through furrr's
future_map function.
}
